bathroom: turn on light for 10 minutes after each change
The Shelly motion sensor does not seem to report continuous events during
motion, but it does seem to reliably report events for motion start and motion
stop, so we now listen to both!

After turning off the light, changes in motion (motion stop after leaving the
room) are ignored for a minute.